init fix and improvement
