+++ ui/src/features/Editor/hooks.ts
@@ -8,14 +8,13 @@ import {
   useState,
 } from "react";
 import { useY } from "react-yjs";
-import { Doc, Array as YArray, UndoManager as YUndoManager } from "yjs";
+import { Array as YArray, UndoManager as YUndoManager } from "yjs";
 
 import { DEFAULT_ENTRY_GRAPH_ID } from "@flow/global-constants";
 import { useShortcuts } from "@flow/hooks";
 import { useSharedProject } from "@flow/lib/gql";
 import { checkForReader } from "@flow/lib/reactFlow";
 import { useYjsStore } from "@flow/lib/yjs";
-import { yWorkflowConstructor } from "@flow/lib/yjs/conversions";
 import type { YWorkflow } from "@flow/lib/yjs/types";
 import useWorkflowTabs from "@flow/lib/yjs/useWorkflowTabs";
 import { useCurrentProject } from "@flow/stores";
@@ -28,28 +27,14 @@ import useNodeLocker from "./useNodeLocker";
 import useUIState from "./useUIState";
 
 export default ({
-  yDoc,
+  yWorkflows,
   undoManager,
   undoTrackerActionWrapper,
 }: {
-  yDoc: Doc;
+  yWorkflows: YArray<YWorkflow>;
   undoManager: YUndoManager | null;
   undoTrackerActionWrapper: (callback: () => void) => void;
 }) => {
-  const yWorkflows = yDoc.getArray<YWorkflow>("workflows");
-
-  useEffect(() => {
-    if (yWorkflows.length === 0) {
-      yDoc.transact(() => {
-        const yWorkflow = yWorkflowConstructor(
-          DEFAULT_ENTRY_GRAPH_ID,
-          "Main Workflow",
-        );
-        yWorkflows.insert(0, [yWorkflow]);
-      });
-    }
-  }, [yWorkflows, yDoc]);
-
   const { fitView } = useReactFlow();
 
   const [currentWorkflowId, setCurrentWorkflowId] = useState<string>(
